requirements.txt
autopep8==2.0.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
distlib==0.3.6
dnspython==2.3.0
fabric==3.0.0
filelock==3.8.0
idna==3.4
invoke==2.0.0
numpy==1.24.2
paho-mqtt==1.6.1
pandas==1.5.3
paramiko==3.0.0
platformdirs==2.5.2
protobuf==4.21.7
pycodestyle==2.10.0
pycparser==2.21
pymongo==4.3.3
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
sentencepiece==0.1.97
six==1.16.0
toml==0.10.2
tomli==2.0.1
torch==1.13.1
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.16.4
